requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
dm-tree==0.1.9
docstring_parser==0.16
einops==0.8.1
etils==1.12.2
executing==2.2.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
flax==0.10.4
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googleapis-common-protos==1.69.2
gradio==5.23.1
gradio_client==1.8.0
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanize==4.12.2
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
namex==0.0.8
narwhals==1.32.0
nest-asyncio==1.6.0
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.11.10
orjson==3.10.16
packaging==23.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
promise==2.3
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
ruff==0.11.2
safehttpx==0.1.6
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
setuptools==77.0.1
shellingham==1.5.4
simple-parsing==0.1.7
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
streamlit==1.44.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-datasets==4.9.8
tensorflow-hub==0.16.1
tensorflow-metadata==1.16.1
tensorflow_decision_forests==1.12.0
tensorflowjs==4.22.0
tensorstore==0.1.72
termcolor==2.5.0
tf_keras==2.19.0
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.9
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wurlitzer==3.1.1
ydf==0.11.0
zipp==3.21.0
